♻️ use temporal app instead of relative folder in `zane_api` app
