+++ backend/temporal/management/commands/run_worker.py
@@ -2,7 +2,7 @@ import asyncio
 
 from django.core.management.base import BaseCommand
 
-from ...temporal.worker import run_worker
+from ...worker import run_worker
 
 
 class Command(BaseCommand):
