--- backend/zane_api/temporal/activities/__init__.py
@@ -1,3 +1,0 @@
-from .service_auto_update import *
-from .main_activities import *
-from .git_activities import *
