--- backend/zane_api/temporal/schedules/__init__.py
@@ -1,2 +1,0 @@
-from .activities import *
-from .workflows import *
