+++ backend/zane_api/tests/base.py
@@ -41,16 +41,16 @@ from ..models import (
     URL,
     Environment,
 )
-from ..temporal.helpers import (
+from temporal.helpers import (
     get_network_resource_name,
     get_env_network_resource_name,
     DockerImageResultFromRegistry,
     SERVER_RESOURCE_LIMIT_COMMAND,
     get_config_resource_name,
 )
-from ..temporal import get_workflows_and_activities
+from temporal.workflows import get_workflows_and_activities
 
-from ..temporal.activities import (
+from temporal.activities import (
     get_swarm_service_name_for_deployment,
     get_volume_resource_name,
 )
@@ -425,7 +425,7 @@ class AuthAPITestCase(APITestCase):
             self.commit_callbacks.append(func)
 
         patch_temporal_client = patch(
-            "zane_api.temporal.main.get_temporalio_client", new_callable=AsyncMock
+            "temporal.main.get_temporalio_client", new_callable=AsyncMock
         )
 
         async def create_schedule(
@@ -459,19 +459,19 @@ class AuthAPITestCase(APITestCase):
                 self.workflow_schedules.remove(schedule_handle)
 
         patch_temporal_create_schedule = patch(
-            "zane_api.temporal.activities.main_activities.create_schedule",
+            "temporal.activities.main_activities.create_schedule",
             side_effect=create_schedule,
         )
         patch_temporal_pause_schedule = patch(
-            "zane_api.temporal.activities.main_activities.pause_schedule",
+            "temporal.activities.main_activities.pause_schedule",
             side_effect=pause_schedule,
         )
         patch_temporal_unpause_schedule = patch(
-            "zane_api.temporal.activities.main_activities.unpause_schedule",
+            "temporal.activities.main_activities.unpause_schedule",
             side_effect=unpause_schedule,
         )
         patch_temporal_delete_schedule = patch(
-            "zane_api.temporal.activities.main_activities.delete_schedule",
+            "temporal.activities.main_activities.delete_schedule",
             side_effect=delete_schedule,
         )
         patch_temporal_create_schedule.start()
