+++ backend/zane_api/views/docker.py
@@ -5,7 +5,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from .. import serializers
-from ..temporal.helpers import (
+from temporal.helpers import (
     search_images_docker_hub,
     check_if_port_is_available_on_host,
 )
