+++ backend/zane_api/views/serializers.py
@@ -1,5 +1,4 @@
 import base64
-from datetime import timedelta
 from io import StringIO
 import json
 import re
@@ -39,7 +38,7 @@ from ..models import (
     DeploymentURL,
     DeploymentChange,
 )
-from ..temporal.helpers import (
+from temporal.helpers import (
     check_if_docker_image_exists,
     check_if_port_is_available_on_host,
     get_server_resource_limits,
